[refactor] : package structure
