+++ src/main/java/com/precapstone/fiveguys_backend/api/redis/RedisService.java
@@ -1,4 +1,4 @@
-package com.precapstone.fiveguys_backend.api.service;
+package com.precapstone.fiveguys_backend.api.redis;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.StringRedisTemplate;
